--- src/main/java/com/daalgae/daalgaeproject/login/model/dao/MemberDAO.java
@@ -1,7 +1,0 @@
-package com.daalgae.daalgaeproject.login.model.dao;
-
-import com.daalgae.daalgaeproject.login.model.dto.MemberDTO;
-
-public interface MemberDAO {
-    MemberDTO findMemberById(String memberId);
-}
