+++ src/main/java/com/daalgae/daalgaeproject/member/controller/LoginController.java
@@ -1,8 +1,7 @@
-package com.daalgae.daalgaeproject.login.controller;
+package com.daalgae.daalgaeproject.member.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
