+++ src/main/java/com/daalgae/daalgaeproject/member/model/dao/MemberDAO.java
@@ -1,0 +1,7 @@
+package com.daalgae.daalgaeproject.member.model.dao;
+
+import com.daalgae.daalgaeproject.member.model.dto.MemberDTO;
+
+public interface MemberDAO {
+    MemberDTO findMemberById(String memberId);
+}
