+++ src/main/java/com/daalgae/daalgaeproject/member/model/dto/AuthenticatedMenuDTO.java
@@ -1,0 +1,16 @@
+package com.daalgae.daalgaeproject.member.model.dto;
+
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+public class AuthenticatedMenuDTO {
+    private int authCode;
+    private int menuCode;
+    private GlobalMenuDTO globalMenu;
+
+}
