+++ src/main/java/com/daalgae/daalgaeproject/member/model/dto/GlobalMenuDTO.java
@@ -1,0 +1,16 @@
+package com.daalgae.daalgaeproject.member.model.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+public class GlobalMenuDTO {
+    private int menuCode;
+    private String menuName;
+    private String menuUrl;
+    private int type;
+    private Integer refMenuCode;
+}
