+++ src/main/java/com/daalgae/daalgaeproject/member/model/dto/MemberRoleDTO.java
@@ -1,0 +1,16 @@
+package com.daalgae.daalgaeproject.member.model.dto;
+
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+public class MemberRoleDTO {
+
+    private int memCode;
+    private int authCode;
+    private AuthorityDTO authority;
+}
