+++ src/main/resources/application.yml
@@ -4,9 +4,13 @@ server:
 spring:
   datasource:
     driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://mysql.coffit.today:3306/daalgae
-    username: Hlkdt6
-    password: Hlkdt6
+    #url: jdbc:mysql://mysql.coffit.today:3306/daalgae
+    #username: Hlkdt6
+    #password: Hlkdt6
+
+    url: jdbc:mysql://localhost:3306/daalgaeSample
+    username: ohgiraffers
+    password: ohgiraffers
 
 mybatis:
   mapper-locations: mappers/**/*.xml
