+++ src/main/resources/mybatis/mapper/MemberDAO.xml
@@ -3,6 +3,6 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.daalgae.daalgaeproject.login.model.dao.MemberDAO">
+<mapper namespace="com.daalgae.daalgaeproject.member.model.dao.MemberDAO">
 
 </mapper>
