+++ src/main/resources/templates/myPage/delete.html
@@ -1,5 +1,6 @@
 <!DOCTYPE html>
-<html lang="ko" dir="ltr">
+<html lang="ko"
+      xmlns:th="http://www.thymeleaf.org">
   <head>
     <meta charset="utf-8">
     <title>회원탈퇴</title>
@@ -11,14 +12,13 @@
 
       
       <!--bootstrap-->
-      <link href="../bootstrap/css/bootstrap.min.css" rel="stylesheet">
+      <link th:href="@{/bootstrap/css/bootstrap.min.css}" rel="stylesheet">
       
 
       <!--css-->
-      <link rel="stylesheet" href="../css/index.css">
-      <link rel="stylesheet" href="../myPageDetails/css/mobile.css">
-      <link rel="stylesheet" href="../css/sanitize.css">
-      <link rel="stylesheet" href="../myPageDetails/css/delete.css">
+      <link rel="stylesheet" th:href="@{/css/index.css}">
+      <link rel="stylesheet" th:href="@{/css/sanitize.css}">
+      <link rel="stylesheet" th:href="@{/css/delete.css}">
       
       <link href="https://unpkg.com/aos@2.3.1/dist/aos.css" rel="stylesheet">
       <link rel="stylesheet" href="http://cdn.jsdelivr.net/npm/xeicon@2.3.3/xeicon.min.css">
@@ -31,10 +31,11 @@
 
       
     <!--favicon--->
-      <link href="favicon.ico" rel="icon">
-      <link href="favicon.ico" rel="apple-touch-icon">
+      <link th:href="@{favicon.ico}" rel="icon">
+      <link th:href="@{favicon.ico}" rel="apple-touch-icon">
   </head>
   <body>
+  <div th:insert="/component/header"></div>
         <section id="delete">
           <div class="deleteWrap container">
 
@@ -51,6 +52,7 @@
             </div>
           </div>
         </section>
+  <div th:insert="/component/footer"></div>
   </body>
